Use ExecutionOptions for new programs
