+++ qiskit_ibm_runtime/options/executor_options.py
@@ -18,26 +18,9 @@ from __future__ import annotations
 from pydantic.dataclasses import dataclass
 from pydantic import Field
 
+from .execution_options import ExecutionOptions
 from .environment_options import EnvironmentOptionsV2
 from .utils import PRIMITIVES_CONFIG
-
-
-@dataclass(config=PRIMITIVES_CONFIG)
-class ExecutionOptions:
-    """Low-level execution options."""
-
-    init_qubits: bool = True
-    """Whether to reset the qubits to the ground state for each shot."""
-
-    rep_delay: float | None = None
-    """The repetition delay.
-
-    This is the delay between a measurement and the subsequent quantum circuit. This is only
-    supported on backends that have ``backend.dynamic_reprate_enabled=True``. It must be from the
-    range supplied by ``backend.rep_delay_range``.
-
-    Default is given by ``backend.default_rep_delay``.
-    """
 
 
 @dataclass(config=PRIMITIVES_CONFIG)
